Issue-16 split app into different components
