+++ src/components/RecipeView.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import './App.css';
+
+class RecipeView extends Component {
+  render() {
+    return (
+      <p className="App-intro">
+        This is where  get started, edit <code>src/App.js</code> and save to reload.
+      </p>
+    );
+  }
+}
+
+export default RecipeView;
