+++ src/components/Sidebar.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import './App.css';
+
+class Sidebar extends Component {
+  render() {
+    return (
+      <div>this is the sidebar</div>
+    );
+  }
+}
+
+export default Sidebar;
